Matching python with new identifiers

+++ papas/graphtools/src/GraphBuilder.cpp
@@ -33,7 +33,7 @@ GraphBuilder::GraphBuilder(const Ids& ids, Edges&& edges) : m_edges(edges), m_el
 
 void GraphBuilder::sortIds(Ids& ids) {
 #if WITHSORT
-  ids.sort(std::greater<int>()); //sort in descending order
+  ids.sort(std::greater<uint64_t>()); //sort in descending order
 #endif
 }
